adjust week2 style
